Reference Depot operators in new non-inner classes.

+++ libraries/whirled/src/java/com/whirled/game/server/persist/GameCookieRepository.java
@@ -12,8 +12,8 @@ import com.google.inject.Singleton;
 import com.samskivert.depot.DepotRepository;
 import com.samskivert.depot.PersistenceContext;
 import com.samskivert.depot.PersistentRecord;
-import com.samskivert.depot.operator.Conditionals;
 import com.samskivert.depot.clause.Where;
+import com.samskivert.depot.operator.In;
 
 /**
  * Provides storage services for user cookies used in games.
@@ -53,8 +53,7 @@ public class GameCookieRepository extends DepotRepository
      */
     public void purgePlayers (Collection<Integer> playerIds)
     {
-        deleteAll(GameCookieRecord.class,
-                  new Where(new Conditionals.In(GameCookieRecord.USER_ID, playerIds)));
+        deleteAll(GameCookieRecord.class, new Where(new In(GameCookieRecord.USER_ID, playerIds)));
     }
 
     @Override // from DepotRepository
